broke mouse events into new components and got click event working.
