+++ src/index.html
@@ -7,7 +7,7 @@
   <title>Solar System</title>
 </head>
 <body>
-  <div class="search"></div>
-  <div id="planet-zone" class='container'></div></div>
+  <div id="search"></div>
+  <div id="planet-zone"></div>
 </body>
 </html>
