+++ src/javascripts/components/planet-cards/planet-cards.scss
@@ -1,3 +1,5 @@
+
+
 #planet-zone {
   display: flex;
   flex-wrap: wrap;
